refactor: abstract timeit func
